rename: RacingCarResult 중복 클래스 삭제 및 이름 변경
